refactor: modularize ClawSweeper and keep fix reviews read-only
